mental-health-support-app/mental-health-support-app/requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
plotly>=5.17.0
requests>=2.31
python-dotenv>=1.0
openai>=1.4.0
